generate the command documents
