+++ docs/commandline/simon_gen-doc.md
@@ -1,0 +1,25 @@
+## simon gen-doc
+
+Generate markdown document for your project
+
+```
+simon gen-doc [flags]
+```
+
+### Options
+
+```
+  -h, --help                      help for gen-doc
+  -d, --output-directory string   assign a directory to store documents (default "./docs/commandline")
+```
+
+### Options inherited from parent commands
+
+```
+      --log-flush-frequency duration   Maximum number of seconds between log flushes (default 5s)
+```
+
+### SEE ALSO
+
+* [simon](simon.md)	 - Simon is a simulator, which will simulate a cluster and simulate workload scheduling.
+
