+++ docs/commandline/simon_version.md
@@ -1,0 +1,24 @@
+## simon version
+
+Print the version of simon
+
+```
+simon version [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for version
+```
+
+### Options inherited from parent commands
+
+```
+      --log-flush-frequency duration   Maximum number of seconds between log flushes (default 5s)
+```
+
+### SEE ALSO
+
+* [simon](simon.md)	 - Simon is a simulator, which will simulate a cluster and simulate workload scheduling.
+
