+++ example/simon-config.yaml
@@ -1,10 +1,24 @@
+#  CR资源说明：由于CR资源行为存在不确定性，open-simulator暂时未支持CR资源。但我们计划制定一种CR资源规范，来消除其不确定性，从而达到支持的目的。
+
 apiVersion: simon/v1alpha1
 kind: Config
 metadata:
   name: simon-config
 spec:
+  # cluster: 导入生成初始集群的配置文件(以下皆为文件路径)
+  #   customConfig: 自定义集群的配置文件
+  #   kubeConfig: 真实集群的kube-config文件
+  #   以上两者取其一
   cluster:
     customConfig: example/cluster/demo_1
+    # kubeConfig:
+
+  # appList: 导入需部署的应用
+  # 支持chart和非chart文件
+  # 多个应用时，部署顺序为配置顺序
+  #   name: 应用名称
+  #   path: 应用文件
+  #   chart: 文件格式可为文件夹或者压缩包格式。若chart指定为true，则表示应用文件为chart文件，若为false或者不指定chart则为非chart文件
   appList:
     - name: yoda
       path: example/application/charts/yoda
@@ -17,4 +31,6 @@ spec:
       path: example/application/open_local
     - name: more_pods
       path: example/application/more_pods
+
+  # newNode: 导入调整集群规模的节点配置文件，节点规格可根据需求任意指定。目前只支持配置一个节点
   newNode: example/node.yaml
